feat(ui): customize main window
